Test a new, very colourful theme.
I have mixed feelings about this new theme. On one hand, the bright
colours have an awesome visual impact without being too powerful. All of
the colours pass WCAG AA (not AAA like with the old theme), so the
visibility should be fine. The pinstriped background is pretty cool, and
I don't think that the animation is too annoying.

On the other hand, the old theme had a beautiful minimalism to it.
Everything was as plain as possible, while still being very precise,
which I do appreciate. I stuck to the OS defaults as much as possible
which makes the website fit in. In addition, a hybrid light/dark theme
is pretty much a must-have these days.

I feel pretty conflicted, so I'm just leaving this new theme as an
experimental branch for now. I've only changed the CSS, with no HTML/JS
changes, so this will be easy enough to merge in later.