Fix svg optimization
`svgo` is breaking `icon-maskable.svg`. There is a specific config
setting that can be disabled to prevent this, but `svgo` v2.0 removes
nearly all CLI options and makes you use an arcane `.js` file for
configuration. Therefore, we need to pin `svgo` to <v2.0 so that we can
adjust the config.