Fix missing newlines in `src/items-to-csv.xslt`.
The GitHub Action `.github/workflows/pages-deploy.yaml` is set to use
`ubuntu-latest`, which was Ubuntu 18.04 for the initial release and
Ubuntu 20.04 as of the last commit 954cc5d (90ff811). For the most recent
build cd7ac80 (bba081f), `ubuntu-latest` points to Ubuntu 22.04.

tdewolff/minify@200330f7 (first in v2.6.0) made it so that `minify`
automatically expands any character entities in the input XML.
Unfortunately, due to XML attribute whitespace normalization, a newline
in an attribute is processed as a space. Ubuntu 20.04 contains
`minify` v2.5.2 while Ubuntu 22.04 contains `minify` v2.9.22, so the
recent upgrade caused the issue to appear.

To fix this, we'll use `<xsl:text>` elements instead of a `$variable`
since elements use less aggressive whitespace normalization than
attributes. We still need to use a character entity plus a trailing
space though to prevent the minifier + normalization combo from deleting
the newline.

Thanks to @WoodlandSkipper for tracking down the issue in the comments
of #5.